% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATORDs_II.R
\name{ATORDs_II}
\alias{ATORDs_II}
\title{Asymmetric Third Order Rotatable Designs based on t-designs}
\usage{
ATORDs_II(v)
}
\arguments{
\item{v}{Number of input factors}
}
\value{
Asymmetric Third Order Rotatable Designs (ATORDs) for a given v.
}
\description{
Generates asymmetric third order rotatable designs based on t- designs of unequal set sizes for a given number of input factors v (4<=v<=9) with coded levels of the factors. Design size is smaller than the design produced by by method 1 in function ATORDs_I.
}
\note{
The user is given with a set of mixed level response surface designs and after entering the serial number of any of the designs from the displayed set, the chosen design will be generated and displayed.
}
\examples{
if(interactive()){
library(MixedLevelRSDs)
}
}
\references{
1)Damaraju Raghavarao & Bei Zhou (1998). < doi: 10.1080/03610929808832657>. "Universal optimality of ue 3-designs for a competing effects model".

2) J.S. Mehta and M.N. Das (1968). "Asymmetric rotatable designs and orthogonal transformations".

3) M. Hemavathi, Shashi Shekhar, Eldho Varghese, Seema Jaggi, Bikas Sinha & Nripes Kumar Mandal (2022)<doi: 10.1080/03610926.2021.1944213>."Theoretical developments in response surface designs: an informative review and further thoughts".
}
